requirements.txt
# usa pip install -r requirements.txt
torch-directml
torchaudio
librosa
numpy
scipy
scikit-learn
pandas
tqdm
matplotlib
soundfile
torchvision
